Expose all models (#2)
This pull request exposes a simple actor for each supported model. The
actual implementation of the models is `LLM`, which is just a
convenience wrapper around MLX's stuff.

Each model conforms to a new protocol called `ModelProtocol`. This
allows us to add extra functions for each model in just a single place:
ModelProtocol.swift. The first example of this is
`request(_:maxTokenCount:)`.

Because of the reentrancy problem with Swift Actors, `ModelProtocol.llm`
is wrapped inside of `ActorLock`, which is taken from [Apple's
swift-build/ASyncLock.swift](https://github.com/swiftlang/swift-build/blob/main/Sources/SWBUtil/AsyncLock.swift).

According to MLX's documentation, the AI models themselves are not
thread-safe, which means calls to them need to be serialized. However,
because Swift Actors are reentrant, calling `try await
llm.request(_:maxTokenCount:)` could immediately suspend and allow
another reference to the same actor be called. This may not be a problem
with library today, but I think it may be in the future, especially when
we add support for `KVCache`. I think it's better to ensure that every
call to `ModelProtocol.someFunc` is transactional, which is what we are
doing by wrapping the implementation of every method on `ModelProtocol`
inside of a `AsyncLock`.