Add processing LLM argument and update mlx-swift-examples (#12)
The processing: UserInput.Processing? argument in LLM.init() was added because it was hoped that, by decreasing the resizing size of images passed to vision models, inference could happen faster and with fewer resources. Unfortunately, at least in the case of Gemma 3 models, the resizing value in UserInput.Processing is ignored in favor of the hard-coded values from the model's configuration. Since other processing options may be added in the future, though, the decision was made to keep the UserInput.Processing argument since it did not change the public interface or behavior of the current implementation at all.